Constructing the maze
Constructing the maze in Unity.

Found methods to build in the map of the maze by two different methods:
1. By using a text file with 1's and 0's.
2. By using Texture Map from an image psd file (note in the future you need to edit the import settings).
Also note that the MapData.cs file takes the text file also from the Resources->Mapdata folder area for the text file that has the same name as the current scene (ie. demo.txt).